Treat javaws as a special case in run-java-tool
We used to have to install dangling symlinks with icedtea(-bin)
pointing to a subsequent installation of icedtea-web. This was
convoluted and didn't make sense as Oracle's VM can also be used
against icedtea-web. Now we always run icedtea-web's javaws if it is
installed. Those wanting Oracle's javaws can simply uninstall
icedtea-web.